requirements.txt
asgiref==3.7.2
Django==4.2.3
sqlparse==0.4.4
typing_extensions==4.7.1
awesome-slugify==1.6.5
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.2.0
django-environ==0.10.0
django-modeltranslation==0.18.10
django-rosetta==0.9.9
django-summernote==0.8.20.0
idna==3.4
mysqlclient==2.2.0
Pillow==10.0.0
polib==1.2.0
protobuf==3.20.3
regex==2023.6.3
requests==2.31.0
six==1.16.0
stripe==5.4.0
Unidecode==0.4.21
urllib3==2.0.3
webencodings==0.5.1
